requirements.txt
aiogram>=3.0.0,<4.0.0
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.0
